chore: moved all collection interfaces for mongo to own dir
