+++ src/lib/server/db/mongo/email-alerts-db-client.ts
@@ -1,7 +1,7 @@
 import type { Collection, Db } from "mongodb"
 import type { IEmailAlertsDbClient } from "$lib/types/db/db-client"
 import type { MetricCount, NewDbEmailAlert } from "$lib/types/db/shared-types"
-import { incrementCount, metricResultFailure, metricResultName, metricResultSuccessful } from "../metrics/handle-metrics"
+import { incrementCount, metricResultFailure, metricResultName, metricResultSuccessful } from "../../metrics/handle-metrics"
 
 export class EmailAlertsDbClient implements IEmailAlertsDbClient {
   private emailAlertsCollection: Collection<NewDbEmailAlert>
